refactor: move quantize to Quantize object
