+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/Module.scala
@@ -20,6 +20,8 @@ import java.nio.ByteOrder
 import com.intel.analytics.bigdl.Module
 import com.intel.analytics.bigdl.nn.abstractnn.Activity
 import com.intel.analytics.bigdl.nn.abstractnn.AbstractModule
+import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity}
+import com.intel.analytics.bigdl.nn.quantization.Quantize
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.File
@@ -149,153 +151,6 @@ object Module {
 
   def quantize[@specialized(Float, Double) T: ClassTag](
     model: AbstractModule[Activity, Activity, T])(implicit ev: TensorNumeric[T]): Module[T] = {
-    type ModuleNode[R] = AbstractModule[Activity, Tensor[R], R]
-    type SeqNodes[R] = Seq[Node[ModuleNode[R]]]
-    type ArrayNodes[R] = Array[Node[ModuleNode[R]]]
-    type ANode[R] = Node[ModuleNode[R]]
-
-    def convertGraph(model: Graph[T]): Graph[T] = {
-      def replaceRef(node: ANode[T], newNode: ANode[T], refs: SeqNodes[T]): Unit = {
-        val buffer = refs.asInstanceOf[ArrayBuffer[Node[ModuleNode[T]]]]
-        refs.zipWithIndex.filter(_._1 == node).foreach { x =>
-          buffer.update(x._2, newNode)
-        }
-      }
-
-      def replace(node: ANode[T], module: ModuleNode[T], list: ArrayNodes[T], index: Int): Unit = {
-        // create a new node
-        val newNode = Node(module)
-        newNode.nextNodes.asInstanceOf[ArrayBuffer[Node[ModuleNode[T]]]] ++= node.nextNodes
-        newNode.prevNodes.asInstanceOf[ArrayBuffer[Node[ModuleNode[T]]]] ++= node.prevNodes
-
-        // prev.next
-        newNode.prevNodes.foreach(n => replaceRef(node, newNode, n.nextNodes))
-
-        // next.prev
-        newNode.nextNodes.foreach(n => replaceRef(node, newNode, n.prevNodes))
-
-        // update the list
-        list.update(index, newNode)
-      }
-
-      val sortedNodes = model.backGraph.topologySort
-
-      for (i <- sortedNodes.indices) {
-        val node = sortedNodes(i)
-        val module = node.element
-        val waitedModule = substitute(module)
-
-        if (waitedModule != module) {
-          replace(node, waitedModule.asInstanceOf[ModuleNode[T]], sortedNodes, i)
-        }
-      }
-
-      val inputs = sortedNodes.filter(n => n.prevNodes.isEmpty)
-      // because all outputs point to dummy nodes, we should filter these nodes as outputs of Graph
-      val outputs = sortedNodes.filter {n =>
-        n.nextNodes.length == 1 && (n.nextNodes.head.element match {
-          case d if d.isInstanceOf[Dummy[T]] => true
-          case _ => false
-        })
-      }
-
-      // create a new Graph, much simpler than replacing others in the old graph
-      Graph(inputs, outputs)
-    }
-
-    def substitute(model: Module[T]): Module[T] = {
-      model match {
-        case container: Container[Activity, Activity, T] =>
-          container match {
-            case graph: Graph[T] =>
-              convertGraph(graph)
-            case _ =>
-              // do with container
-              for (i <- container.modules.indices) {
-                container.modules(i) = substitute(container.modules(i))
-              }
-              container
-          }
-        case normalConv if normalConv.isInstanceOf[SpatialConvolution[T]] =>
-          // do with normal convolution
-          val conv = normalConv.asInstanceOf[SpatialConvolution[T]]
-          val quantizedConv = new quantization.SpatialConvolution[T](
-            conv.nInputPlane,
-            conv.nOutputPlane,
-            conv.kernelW,
-            conv.kernelH,
-            conv.strideW,
-            conv.strideH,
-            conv.padW,
-            conv.padH,
-            conv.nGroup)
-          quantizedConv.initWeightAndBias(conv.weight, conv.bias)
-        case dilatedConv if dilatedConv.isInstanceOf[SpatialDilatedConvolution[T]] =>
-          // do with dilated convolution
-          val conv = dilatedConv.asInstanceOf[SpatialDilatedConvolution[T]]
-          val quantizedConv = new quantization.SpatialDilatedConvolution[T](
-            conv.nInputPlane,
-            conv.nOutputPlane,
-            conv.kW,
-            conv.kH,
-            conv.dW,
-            conv.dH,
-            conv.padW,
-            conv.padH,
-            conv.dilationW,
-            conv.dilationH)
-          quantizedConv.initWeightAndBias(conv.weight, conv.bias)
-        case normalLinear if normalLinear.isInstanceOf[Linear[T]] =>
-          // do with linear
-          val linear = normalLinear.asInstanceOf[Linear[T]]
-
-          val quantizedLinear = new quantization.Linear[T](
-            linear.weight.size(2),
-            linear.weight.size(1))
-
-          quantizedLinear.initWeightAndBias(linear.weight, linear.bias)
-        case _ => model
-      }
-    }
-
-    def reorganizeParameters(parameters: Array[Tensor[T]]): Tensor[T] = {
-      var length = 0
-      for (i <- parameters.indices) {
-        if (parameters(i) != null) {
-          length += parameters(i).nElement()
-        }
-      }
-
-      val result = Tensor[T](length)
-
-      var offset = 0
-      for (i <- parameters.indices) {
-        val parameter = parameters(i)
-
-        if (parameter != null) {
-          val length = parameter.nElement()
-
-          val (src, srcOffset) = (parameter.storage().array(), parameter.storageOffset() - 1)
-          val (dst, dstOffset) = (result.storage().array(), offset)
-
-          val (size, stride) = (parameter.size(), parameter.stride())
-
-          System.arraycopy(src, srcOffset, dst, dstOffset, length)
-          parameter.set(result.storage(), offset + 1, size, stride)
-
-          offset += length
-        }
-      }
-
-      result
-    }
-
-    // deep copy a new model then substitute with all quantized version models
-    val quantizedModel = substitute(model.cloneModule())
-
-    val paras = quantizedModel.parameters()._1
-    reorganizeParameters(paras)
-
-    quantizedModel
+    Quantize.quantize(model)
   }
 }
